requirements/requirements.txt
datasets==2.17.0
numpy>=1.26.4
pydantic>=1.8.2
pytest>=7.2.1
pytorch_lightning==2.0.8
PyYAML==6.0.1
setuptools==60.5.0
strictyaml==1.7.3
tokenizers==0.13.3
